moved IterableWeakMap to utils
